#426 Eliminate warnings
Removed one unnecessary cast to int, and refactored a method to remove
an unused parameter, which required changing its method references to
lambda expressions.

+++ src/ca/mcgill/cs/jetuml/views/StringViewer.java
@@ -259,7 +259,7 @@ public final class StringViewer
 			if( aAlign.isHorizontallyCentered() )
 			{
 				xOffset = dimension.width()/2;
-				yOffset = (int) (CANVAS_FONT.fontSize()/2) + 1;
+				yOffset = CANVAS_FONT.fontSize()/2 + 1;
 			}
 			else if( aAlign.isRight() )
 			{
